Changes in lib
