--- google/src/main/java/com/google/callback/OnCancel.kt
@@ -1,5 +1,0 @@
-package com.google.callback
-
-interface OnCancel {
-    fun onCancel()
-}
