--- google/src/main/java/com/google/callback/OnError.kt
@@ -1,5 +1,0 @@
-package com.google.callback
-
-interface OnError<ERROR> {
-    fun onError(error: ERROR)
-}
