--- google/src/main/java/com/google/callback/OnSuccess.kt
@@ -1,5 +1,0 @@
-package com.google.callback
-
-interface OnSuccess<RESULT> {
-    fun onSuccess(result: RESULT)
-}
