+++ readMe.md
@@ -6,47 +6,56 @@ Go to -> [Firebase Console](https://console.firebase.google.com/)
 Enable `Google` from `Authentication` `SignIn` and copy webclientId
 
 ```kotlin
-GoogleConfiguration.clientId(webclientId)
-                .build();
+config("clientId")
 ```
 
 To check for current user
 
 ```kotlin
-GoogleConfiguration.getAuth().getCurrentUser();
+GoogleManager.user
 ```
 
 `Login Sample`
 
 ```kotlin
-FirebaseUser user = GoogleConfiguration.getAuth().getCurrentUser();
-            if (user == null) {
-                GoogleConnect.INSTANCE.with()
-                        .login(activity, requestCode)
-                        .build();
-            } else {
-                Log.i(getLocalClassName(), user.getDisplayName() + " " + user.getEmail() + "" + user.getPhoneNumber());
-            }
+login(1000)
 ```
 
 `Profile Sample`
 
 ```kotlin
-GoogleConnect.with()
-             .profile()
-             .build()
+profile({
+                            Log.i(
+                                localClassName + "Google",
+                                "${it.displayName} ${it.email} ${it.phoneNumber}"
+                            )
+                        }, {
+                            it.printStackTrace()
+                        }, {
+                            Log.i(
+                                localClassName ,"Cancelled"
+                            )
+                        })
 ```
 
 `OnAcitivyResult`
 
 ```kotlin
-GoogleConfiguration.onActivityResult(data, aBoolean -> {
-                    if (aBoolean)
-                        GoogleConnect.INSTANCE.with()
-                                .profile()
-                                .build();
-                    return Unit.INSTANCE;
-                });
+GoogleManager.onActivityResult(data!!) {
+                    if (this)
+                        profile({
+                            Log.i(
+                                localClassName + "Google",
+                                "${it.displayName} ${it.email} ${it.phoneNumber}"
+                            )
+                        }, {
+                            it.printStackTrace()
+                        }, {
+                            Log.i(
+                                localClassName ,"Cancelled"
+                            )
+                        })
+                }
 ```
 
 Download
@@ -64,5 +73,5 @@ repositories {
 ```
 
 ```groovy
-implementation 'com.amitsahni:google:0.0.1-alpha03'
+implementation 'com.amitsahni:google:0.0.1-alpha04'
 ```
